#129 Refactor AgentFlow to move unrelated concerns to other components
First up: Edit Dock.
